+++ WeatherStatistics/WeatherStatistics/Models/MonthStatistic.swift
@@ -12,11 +12,11 @@ class MonthStatistic {
     var month: Int?
     var tmax: Float?
     var tmin: Float?
-    var days: Int?
+    var days: Float?
     var rainMM: Float?
     var sunH: Float?
     
-    init(month: Int?, tmax: Float?, tmin: Float?, days: Int?, rainMM: Float?, sunH: Float?) {
+    init(month: Int?, tmax: Float?, tmin: Float?, days: Float?, rainMM: Float?, sunH: Float?) {
         self.month = month
         self.tmax = tmax
         self.tmin = tmin
